Created helper class for editItemAdapter and extracted viewholders from adapter to separate classes.
