+++ app/src/main/java/com/invisibleteam/goinvisible/mvvm/edition/adapter/EditCompoundRecyclerView.java
@@ -6,20 +6,21 @@ import android.support.annotation.VisibleForTesting;
 import android.support.v7.widget.LinearLayoutManager;
 import android.util.AttributeSet;
 
+import com.invisibleteam.goinvisible.helper.EditItemAdapterHelper;
 import com.invisibleteam.goinvisible.model.Tag;
 import com.invisibleteam.goinvisible.mvvm.common.CompoundRecyclerView;
 import com.invisibleteam.goinvisible.mvvm.edition.OnTagActionListener;
 
 import java.util.List;
 
-public class EditCompoundRecyclerView extends CompoundRecyclerView<Tag, EditItemAdapter.ViewHolder> {
+public class EditCompoundRecyclerView extends CompoundRecyclerView<Tag, ViewHolder> {
 
     private EditItemAdapter itemAdapter;
 
     public EditCompoundRecyclerView(Context context, AttributeSet attrs) {
         super(context, attrs);
 
-        itemAdapter = new EditItemAdapter();
+        itemAdapter = new EditItemAdapter(new EditItemAdapterHelper());
         init(itemAdapter, new LinearLayoutManager(getContext()));
     }
 
@@ -47,7 +48,7 @@ public class EditCompoundRecyclerView extends CompoundRecyclerView<Tag, EditItem
     }
 
     public List<Tag> getAllTags() {
-        return itemAdapter.getAllTags();
+        return itemAdapter.getTagsList();
     }
 
     public void updateTagListAfterChanges(List<Tag> tags) {
