+++ app/src/main/java/com/invisibleteam/goinvisible/mvvm/edition/adapter/EditItemAdapter.java
@@ -8,13 +8,9 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.invisibleteam.goinvisible.R;
-import com.invisibleteam.goinvisible.databinding.EditItemViewBinding;
-import com.invisibleteam.goinvisible.databinding.SectionEditItemViewBinding;
-import com.invisibleteam.goinvisible.model.InputType;
+import com.invisibleteam.goinvisible.helper.EditItemAdapterHelper;
 import com.invisibleteam.goinvisible.model.Tag;
-import com.invisibleteam.goinvisible.model.TagType;
 import com.invisibleteam.goinvisible.mvvm.edition.OnTagActionListener;
-import com.invisibleteam.goinvisible.util.TextViewUtil;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -24,28 +20,20 @@ import static com.invisibleteam.goinvisible.model.TagGroupType.DEVICE_INFO;
 import static com.invisibleteam.goinvisible.model.TagGroupType.IMAGE_INFO;
 import static com.invisibleteam.goinvisible.model.TagGroupType.LOCATION_INFO;
 
-class EditItemAdapter extends RecyclerView.Adapter<EditItemAdapter.ViewHolder> {
+class EditItemAdapter extends RecyclerView.Adapter<ViewHolder> {
 
     private static final int IMAGE_INFO_SECTION_POSITION = 0;
     private static final int LOCATION_INFO_SECTION_POSITION = 21;
     private static final int DEVICE_INFO_SECTION_POSITION = 47;
     private static final int ADVANCED_INFO_SECTION_POSITION = 59;
-    private List<Tag> tagsList;
+
     private List<Tag> baseTagsList;
     private OnTagActionListener onTagActionListener;
     private Context context;
+    private EditItemAdapterHelper helper;
 
-    private List<Tag> imageInfoTagList;
-    private List<Tag> locationInfoTagList;
-    private List<Tag> deviceInfoTagList;
-    private List<Tag> advancedInfoTagList;
-
-    EditItemAdapter() {
-        tagsList = new ArrayList<>();
-        imageInfoTagList = new ArrayList<>();
-        locationInfoTagList = new ArrayList<>();
-        deviceInfoTagList = new ArrayList<>();
-        advancedInfoTagList = new ArrayList<>();
+    EditItemAdapter(EditItemAdapterHelper helper) {
+        this.helper = helper;
     }
 
     @Override
@@ -58,12 +46,12 @@ class EditItemAdapter extends RecyclerView.Adapter<EditItemAdapter.ViewHolder> {
         }
 
         final View itemView = LayoutInflater.from(parent.getContext()).inflate(R.layout.edit_item_view, parent, false);
-        return new TagViewHolder(itemView);
+        return new TagViewHolder(itemView, onTagActionListener);
     }
 
     @Override
     public void onBindViewHolder(ViewHolder holder, int position) {
-        final Tag tag = tagsList.get(position);
+        final Tag tag = getTagsList().get(position);
         holder.itemView.setTag(tag);
 
         if (holder instanceof SectionViewHolder) {
@@ -91,11 +79,11 @@ class EditItemAdapter extends RecyclerView.Adapter<EditItemAdapter.ViewHolder> {
 
     @Override
     public int getItemCount() {
-        return tagsList.size();
+        return getTagsList().size();
     }
 
     void updateTagList(List<Tag> tags) {
-        createGroups(tags);
+        helper.createGroups(tags);
         if (baseTagsList == null) {
             baseTagsList = new ArrayList<>();
             updateBaseTagList(tags);
@@ -104,41 +92,6 @@ class EditItemAdapter extends RecyclerView.Adapter<EditItemAdapter.ViewHolder> {
         }
     }
 
-    private void createGroups(List<Tag> tags) {
-        clearTagLists();
-        addHeadersToLists();
-        for (Tag tag : tags) {
-            if (IMAGE_INFO.equals(tag.getTagGroupType())) {
-                imageInfoTagList.add(tag);
-            } else if (LOCATION_INFO.equals(tag.getTagGroupType())) {
-                locationInfoTagList.add(tag);
-            } else if (DEVICE_INFO.equals(tag.getTagGroupType())) {
-                deviceInfoTagList.add(tag);
-            } else {
-                advancedInfoTagList.add(tag);
-            }
-        }
-
-        this.tagsList.clear();
-        this.tagsList.addAll(imageInfoTagList);
-        this.tagsList.addAll(locationInfoTagList);
-        this.tagsList.addAll(deviceInfoTagList);
-        this.tagsList.addAll(advancedInfoTagList);
-    }
-
-    private void addHeadersToLists() {
-        imageInfoTagList.add(new Tag("", "", TagType.build(InputType.INDEFINITE), IMAGE_INFO)); //todo generate method for this
-        locationInfoTagList.add(new Tag("", "", TagType.build(InputType.INDEFINITE), LOCATION_INFO));
-        deviceInfoTagList.add(new Tag("", "", TagType.build(InputType.INDEFINITE), DEVICE_INFO));
-        advancedInfoTagList.add(new Tag("", "", TagType.build(InputType.INDEFINITE), ADVANCED));
-    }
-
-    private void clearTagLists() {
-        imageInfoTagList.clear();
-        locationInfoTagList.clear();
-        deviceInfoTagList.clear();
-        advancedInfoTagList.clear();
-    }
 
     void updateBaseTagList(List<Tag> tags) {
         baseTagsList.clear();
@@ -149,7 +102,7 @@ class EditItemAdapter extends RecyclerView.Adapter<EditItemAdapter.ViewHolder> {
 
     List<Tag> getChangedTags() {
         List<Tag> changedTags = new ArrayList<>();
-        for (Tag tag : tagsList) {
+        for (Tag tag : getTagsList()) {
             for (Tag baseTag : baseTagsList) {
                 if (isTagChanged(baseTag, tag)) {
                     changedTags.add(tag);
@@ -158,10 +111,6 @@ class EditItemAdapter extends RecyclerView.Adapter<EditItemAdapter.ViewHolder> {
             }
         }
         return changedTags;
-    }
-
-    List<Tag> getAllTags() {
-        return tagsList;
     }
 
     private boolean isTagChanged(Tag baseTag, Tag tag) {
@@ -183,10 +132,11 @@ class EditItemAdapter extends RecyclerView.Adapter<EditItemAdapter.ViewHolder> {
     }
 
     void updateTag(Tag tag) {
-        if (!tagsList.isEmpty()) {
-            for (int index = 0; index < tagsList.size(); index++) {
-                if (tagsList.get(index).getKey().equals(tag.getKey())) {
-                    tagsList.set(index, tag);
+        List<Tag> tagList = getTagsList();
+        if (!tagList.isEmpty()) {
+            for (int index = 0; index < tagList.size(); index++) {
+                if (tagList.get(index).getKey().equals(tag.getKey())) {
+                    tagList.set(index, tag);
                     notifyItemChanged(index);
                     onTagActionListener.onTagsUpdated();
                     break;
@@ -197,68 +147,6 @@ class EditItemAdapter extends RecyclerView.Adapter<EditItemAdapter.ViewHolder> {
 
     @VisibleForTesting
     List<Tag> getTagsList() {
-        return tagsList;
-    }
-
-
-    class ViewHolder extends RecyclerView.ViewHolder {
-
-        ViewHolder(View itemView) {
-            super(itemView);
-        }
-    }
-
-    class TagViewHolder extends ViewHolder {
-
-        @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
-        EditItemViewModel editItemViewModel;
-
-        EditItemViewBinding editItemViewBinding;
-
-        TagViewHolder(View itemView) {
-            super(itemView);
-
-            editItemViewModel = new EditItemViewModel();
-            editItemViewBinding = EditItemViewBinding.bind(itemView);
-            editItemViewBinding.setViewModel(editItemViewModel);
-            TextViewUtil.setEllipsizedForView(editItemViewBinding.tagKey);
-            setupListeners();
-        }
-
-        private void setupListeners() {
-            editItemViewBinding.clearButton.setOnClickListener(v -> {
-                if (onTagActionListener != null) {
-                    Tag tag = (Tag) itemView.getTag();
-                    onTagActionListener.onClear(tag);
-                }
-            });
-
-            itemView.setOnClickListener(view -> {
-                if (onTagActionListener != null) {
-                    Tag tag = (Tag) itemView.getTag();
-                    onTagActionListener.onEditStarted(tag);
-                }
-            });
-        }
-    }
-
-    class SectionViewHolder extends ViewHolder {
-
-        @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
-        SectionEditItemViewModel sectionEditItemViewModel;
-
-        SectionEditItemViewBinding sectionEditItemViewBinding;
-
-        SectionViewHolder(View itemView) {
-            super(itemView);
-
-            sectionEditItemViewModel = new SectionEditItemViewModel();
-            sectionEditItemViewBinding = SectionEditItemViewBinding.bind(itemView);
-            sectionEditItemViewBinding.setViewModel(sectionEditItemViewModel);
-        }
-
-        void setSectionName(String name) {
-            sectionEditItemViewModel.setSectionName(name);
-        }
+        return helper.getTagsList();
     }
 }
