+++ app/src/main/java/com/invisibleteam/goinvisible/mvvm/edition/adapter/ViewHolder.java
@@ -1,0 +1,11 @@
+package com.invisibleteam.goinvisible.mvvm.edition.adapter;
+
+import android.support.v7.widget.RecyclerView;
+import android.view.View;
+
+class ViewHolder extends RecyclerView.ViewHolder {
+
+    ViewHolder(View itemView) {
+        super(itemView);
+    }
+}
